+++ source/deprecated/omni.isaac.asset_browser/omni/isaac/asset_browser/__init__.py
@@ -16,5 +16,3 @@ new_extension_name = "isaacsim.asset.browser"
 carb.log_warn(
     f"{old_extension_name} has been deprecated in favor of {new_extension_name}. Please update your code accordingly."
 )
-
-from isaacsim.asset.browser import *