+++ source/extensions/isaacsim.storage.native/isaacsim/storage/native/tests/__init__.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2018-2023, NVIDIA CORPORATION. All rights reserved.
+# Copyright (c) 2018-2024, NVIDIA CORPORATION. All rights reserved.
 #
 # NVIDIA CORPORATION and its licensors retain all intellectual property
 # and proprietary rights in and to this software, related documentation
@@ -6,12 +6,5 @@
 # distribution of this software and related documentation without an express
 # license agreement from NVIDIA CORPORATION is strictly prohibited.
 #
-import carb
 
-old_extension_name = "omni.isaac.assets_check"
-new_extension_name = "isaacsim.asset.browser"
-
-# Provide deprecation warning to user
-carb.log_warn(
-    f"{old_extension_name} has been deprecated in favor of {new_extension_name}. Please update your code accordingly."
-)
+scan_for_test_modules = True