+++ source/extensions/isaacsim.util.internal/isaacsim/util/internal/extension.py
@@ -31,7 +31,7 @@ class InternalTools(omni.ext.IExt):
         self._window = ui.Window(
             title=EXTENSION_NAME, width=800, height=400, visible=False, dockPreference=ui.DockPreference.LEFT_BOTTOM
         )
-        default_server = carb.settings.get_settings().get("/persistent/isaac/asset_root/isaac")
+        default_server = carb.settings.get_settings().get("/persistent/isaac/asset_root/default")
         with self._window.frame:
             with ui.VStack(height=0):
                 with ui.HStack(height=0):