+++ source/extensions/isaacsim.app.setup/docs/CHANGELOG.md
@@ -1,8 +1,11 @@
 # Changelog
+## [1.0.2] - 2024-11-01
+### Changed
+- Make Isaac Sim Assets Browser in focus by default
+
 ## [1.0.1] - 2024-10-24
 ### Changed
 - Updated dependencies and imports after renaming
-
 
 ## [1.0.0] - 2024-10-11
 ### Changed